# [SPARK-38506][SQL] Push partial aggregation through join