@wangyum 2022-03-30
q54 is slower than before because it can't reduce lots of data, but we can improve it by https://issues.apache.org/jira/browse/SPARK-38505

Before | After
-- | --
![image](https://user-images.githubusercontent.com/5399861/160748009-7d1913a8-3c46-4690-ab3c-d38dc4a59eaf.png) | ![image](https://user-images.githubusercontent.com/5399861/160748691-99e46476-3958-4eaf-b84b-8ae39577b8e0.png)