@wangyum 2022-04-14
Here's a real case from production:
Before | After
-- | --
![image](https://user-images.githubusercontent.com/5399861/163328367-b7a94082-5b3c-4d4f-8ae2-0437aa3a2f9c.png)| ![image](https://user-images.githubusercontent.com/5399861/163328588-dcbc86ac-3792-4f5b-bc25-9841d5c8736b.png)