@wangyum 2022-04-14
cc @cloud-fan @sigmod This feature can significantly improve sql performance.